splash page and finished styling forms
