+++ src/pages/app.js
@@ -4,11 +4,13 @@ import Layout from "../components/layout"
 import PrivateRoute from "../components/privateRoute"
 import SecondPage from "./game"
 import Signup from "./signup"
+import Login from "./login"
 
 const App = () => (
   <Layout>
     <Router>
       <Signup path="/app/signup" />
+      <Login path="/app/login" />
       <PrivateRoute path="/game/" component={SecondPage} />
     </Router>
   </Layout>
